Refactor agent example code to use explicit API key configuration and streamline main functions
