+++ src/examples/simple-tools/src/commonMain/kotlin/community/flock/aigentic/example/ImageExample.kt
@@ -2,9 +2,11 @@ package community.flock.aigentic.example
 
 import community.flock.aigentic.core.agent.Run
 import community.flock.aigentic.core.agent.start
-import community.flock.aigentic.core.dsl.AgentConfig
 import community.flock.aigentic.core.dsl.agent
 import community.flock.aigentic.core.message.MimeType
+import community.flock.aigentic.core.model.ModelIdentifier
+import community.flock.aigentic.openai.dsl.openAIModel
+import community.flock.aigentic.openai.model.OpenAIModelIdentifier
 import community.flock.aigentic.core.tool.Parameter
 import community.flock.aigentic.core.tool.ParameterType.Primitive
 import community.flock.aigentic.core.tool.Tool
@@ -35,11 +37,14 @@ val saveItemTool =
 
 suspend fun runItemCategorizeExample(
     base64Image: String,
-    configureModel: AgentConfig.() -> Unit,
-): Run {
-    val run =
+    apiKey: String,
+) {
+    val run: Run =
         agent {
-            configureModel()
+            openAIModel {
+                apiKey(apiKey)
+                modelIdentifier(OpenAIModelIdentifier.GPT4OMini)
+            }
             task("Identify all items in the image and save each individual item") {}
             addTool(saveItemTool)
             context {
@@ -47,5 +52,9 @@ suspend fun runItemCategorizeExample(
             }
         }.start()
 
-    return run
+    when (val result = run.result) {
+        is community.flock.aigentic.core.agent.tool.Result.Finished -> "Agent finished successfully"
+        is community.flock.aigentic.core.agent.tool.Result.Stuck -> "Agent is stuck and could not complete task, it says: ${result.reason}"
+        is community.flock.aigentic.core.agent.tool.Result.Fatal -> "Agent crashed: ${result.message}"
+    }.also(::println)
 }
