+++ src/examples/simple-tools/src/commonMain/kotlin/community/flock/aigentic/example/PdfExtractor.kt
@@ -2,8 +2,9 @@ package community.flock.aigentic.example
 
 import community.flock.aigentic.core.agent.Run
 import community.flock.aigentic.core.agent.start
-import community.flock.aigentic.core.dsl.AgentConfig
 import community.flock.aigentic.core.dsl.agent
+import community.flock.aigentic.gemini.dsl.geminiModel
+import community.flock.aigentic.gemini.model.GeminiModelIdentifier
 import community.flock.aigentic.core.message.MimeType
 import community.flock.aigentic.core.tool.Parameter
 import community.flock.aigentic.core.tool.ParameterType.Primitive
@@ -63,11 +64,14 @@ data class InvoiceComponent(
 
 suspend fun invoiceExtractorAgent(
     invoicePdfBase64: String,
-    configureModel: AgentConfig.() -> Unit,
-): Run {
-    val run =
+    apiKey: String,
+) {
+    val run: Run =
         agent {
-            configureModel()
+            geminiModel {
+                apiKey(apiKey)
+                modelIdentifier(GeminiModelIdentifier.Gemini2_0Flash)
+            }
             task("Extract the different invoice components") {
                 addInstruction("Please provide list of the invoice components")
             }
@@ -77,5 +81,9 @@ suspend fun invoiceExtractorAgent(
             }
         }.start()
 
-    return run
+    when (val result = run.result) {
+        is community.flock.aigentic.core.agent.tool.Result.Finished -> "Agent finished successfully"
+        is community.flock.aigentic.core.agent.tool.Result.Stuck -> "Agent is stuck and could not complete task, it says: ${result.reason}"
+        is community.flock.aigentic.core.agent.tool.Result.Fatal -> "Agent crashed: ${result.message}"
+    }.also(::println)
 }
